requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-huggingface
langchain-pinecone
langchain-text-splitters
pinecone
python-dotenv
streamlit
pymupdf
sentence-transformers
hypercorn==0.14.4
